+++ packages/components/atoms/f-filter-pill/src/components/_tests/FilterPill.test.js
@@ -41,6 +41,23 @@ describe('FilterPill', () => {
             });
         });
 
+        describe('href :: ', () => {
+            it('should have filter `href` value', () => {
+                // Arrange
+                const propsData = { href: '/area/cf10-cardiff/?refine=low_delivery_fee' };
+
+                // Act
+                const wrapper = shallowMount(FilterPill, {
+                    propsData
+                });
+
+                const link = wrapper.find('[data-test-id="filter-pill-link"]');
+
+                // Assert
+                expect(link.attributes('href')).toBe(propsData.href);
+            });
+        });
+
         describe('isSelected :: ', () => {
             it.each([
                 [true, true],
@@ -67,35 +84,6 @@ describe('FilterPill', () => {
 
                 // Assert
                 expect(wrapper.vm.isToggleSelected).toBe(true);
-            });
-        });
-
-        describe('isDisabled :: ', () => {
-            it.each([
-                [true, true],
-                [false, false]
-            ])('should update `isToggleDisabled` %s when set to %s', (expectedValue, isDisabled) => {
-                // Arrange
-                const propsData = { isDisabled };
-
-                // Act
-                const wrapper = shallowMount(FilterPill, {
-                    propsData
-                });
-
-                // Assert
-                expect(wrapper.vm.isToggleDisabled).toBe(expectedValue);
-            });
-
-            it('should update `isToggleDisabled` when changed', async () => {
-                // Arrange
-                const wrapper = shallowMount(FilterPill, { isToggleDisabled: false, isDisabled: false });
-
-                // Act
-                await wrapper.vm.$options.watch.isDisabled.call(wrapper.vm, true);
-
-                // Assert
-                expect(wrapper.vm.isToggleDisabled).toBe(true);
             });
         });