+++ packages/components/atoms/f-filter-pill/test/accessibility/axe-accessibility.spec.js
@@ -1,4 +1,4 @@
-import { getAccessibilityTestResults } from '../../../../../../test/utils/axe-helper'; // eslint-disable-line import/no-relative-packages
+import { getAxeResults } from '../../../../../../test/utils/axe-helper'; // eslint-disable-line import/no-relative-packages
 
 const FilterPill = require('../../test-utils/component-objects/f-filter-pill.component');
 
@@ -11,7 +11,7 @@ describe('Accessibility tests', () => {
 
     it('a11y - should test f-filter-pill component WCAG compliance', () => {
         // Act
-        const axeResults = getAccessibilityTestResults('f-filter-pill');
+        const axeResults = getAxeResults('f-filter-pill');
 
         // Assert
         expect(axeResults.violations.length).toBe(0);