+++ packages/components/molecules/f-restaurant-card/src/components/subcomponents/RestaurantRating/RestaurantRating.vue
@@ -55,22 +55,30 @@
                 {{ ratingsMax }}
             </data>
 
-            <!-- Number of ratings/Own rating message -->
-            <span aria-hidden="true">
+            <!-- Own rating message -->
+            <span
+                v-if="isOwnRating"
+                aria-hidden="true">
+                &#40;
+                <span
+                    data-test-id="rating-own-rating-message"
+                    :class="[$style['c-restaurantCard-rating-count']]">
+                    {{ isOwnRatingMessage }}
+                </span>
+                &#41;
+            </span>
+
+            <!-- Number of ratings -->
+            <span
+                v-else-if="count"
+                aria-hidden="true">
                 &#40;
                 <data
-                    v-if="!isOwnRating"
                     data-test-id="rating"
                     :class="[$style['c-restaurantCard-rating-count']]"
                     :value="count">
                     {{ count }}
                 </data>
-                <span
-                    v-else
-                    data-test-id="rating-own-rating-message"
-                    :class="[$style['c-restaurantCard-rating-count']]">
-                    {{ isOwnRatingMessage }}
-                </span>
                 &#41;
             </span>
         </template>
@@ -146,7 +154,7 @@ export default {
     },
     computed: {
         noRatingsAvailable () {
-            return !this.count || !this.mean;
+            return !this.mean;
         },
         meanFormatted () {
             return Number.parseFloat(this.mean).toFixed(2);