+++ packages/components/organisms/f-header/src/components/UserNavigationPanel.vue
@@ -1,12 +1,14 @@
 <template>
     <ul
         :aria-label="copy.navTitle"
-        :class="$style['c-nav-popoverList']">
+        :class="[$style['c-nav-popoverList'],
+                 $style['c-user-list']]">
         <li
             v-for="(link, index) in copy.navLinks"
             :key="index"
             data-test-id="nav-links">
             <a
+                :class="$style['list-link']"
                 :tabindex="tabIndex"
                 :href="link.url"
                 :data-trak='`{
@@ -15,7 +17,6 @@
                     "action": "header",
                     "label": "${link.gtm}"
                 }`'
-                :class="$style['list-link']"
                 @blur="$emit('deactivateNav')"
                 @focus="$emit('activateNav')">
                 {{ link.text }}
@@ -88,32 +89,32 @@ export default {
 <style lang="scss" module>
 @import '../assets/scss/navigation.scss';
 
+.c-user-list {
+    @include media('>mid') {
+        margin: spacing(b) 0;
+    }
+}
+
 .list-link {
-    display: block;
-    padding: spacing(c) spacing(d);
-    margin: 0;
-    font-family: $font-family-base;
-    font-weight: 300;
-    color: $color-content-subdued;
-    height: auto;
-    @include font-size('body-s');
     text-decoration: none;
-    border-bottom: 1px solid $color-border-default;
+
+    @include media('>mid') {
+        display: block;
+        padding: spacing(c) 0;
+        height: auto;
+    }
+
+    @include media('<=mid') {
+        display: flex;
+        margin-left: spacing(h);
+        padding: spacing(c) spacing(d) spacing(c) 0;
+        border-top: 1px solid $color-border-default;
+    }
 
     &:hover,
     &:focus {
         font-weight: $font-weight-bold;
         text-decoration: none;
-        color: $color-content-subdued;
     }
-
-    @include media('>mid') {
-        @include font-size('body-l');
-    }
-
-    @include media('<=mid') {
-        padding-left: spacing(h);
-    }
-
 }
 </style>