+++ packages/components/pages/f-registration/test/accessibility/axe-accessibility.spec.js
@@ -1,4 +1,4 @@
-const { getAccessibilityTestResults } = require('../../../../../../test/utils/axe-helper');
+const { getAxeResults } = require('../../../../../../test/utils/axe-helper');
 const Registration = require('../../test-utils/component-objects/f-registration.component');
 
 const registration = new Registration();
@@ -11,7 +11,7 @@ describe('Accessibility tests', () => {
 
     it('a11y - should test f-registration component WCAG compliance', () => {
         // Act
-        const axeResults = getAccessibilityTestResults('f-registration');
+        const axeResults = getAxeResults('f-registration');
 
         // Assert
         expect(axeResults.violations.length).toBe(0);